Refactoring of apis folder divided in scim and gnap. Extract handleErrorResponse() in its own file valid for Groups and Users. Add type ErrorDetail to handleErrorResponse parameter that explicity accepts only jsonResponse as input. Cleaned unused types and added ListResponse from typescript-clients/scim folder
