+++ src/apis/scim/usersRequest.ts
@@ -1,8 +1,9 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { AppDispatch, AppRootState } from "init-app";
 import { UserResponse } from "typescript-clients/scim";
-import { fakeEPPN } from "../common/testEPPNData";
-import { baseURL, handleErrorResponse, scimHeaders } from "./scimGroupsRequest";
+import { fakeEPPN } from "../../common/testEPPNData";
+import { handleErrorResponse } from "./error";
+import { baseURL, scimHeaders } from "./groupsRequest";
 
 export const postUser = createAsyncThunk<
   UserResponse, // return type
