+++ src/components/GroupManagement.tsx
@@ -7,8 +7,8 @@ import { Field, Form } from "react-final-form";
 import { FormattedMessage } from "react-intl";
 import { useLocation, useNavigate } from "react-router-dom";
 import { GroupMember } from "typescript-clients/scim/models/GroupMember";
-import { createGroup, getGroupDetails, getGroupsSearch, putGroup } from "../apis/scimGroupsRequest";
-import { getUserDetails, postUser } from "../apis/scimUsersRequest";
+import { createGroup, getGroupDetails, getGroupsSearch, putGroup } from "../apis/scim/groupsRequest";
+import { getUserDetails, postUser } from "../apis/scim/usersRequest";
 import { useAppDispatch, useAppSelector } from "../hooks";
 import appSlice from "../slices/appReducers";
 import getGroupsSlice from "../slices/getGroups";
