+++ src/components/MembersList.tsx
@@ -4,8 +4,8 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Fragment, useEffect, useState } from "react";
 import { FormattedMessage } from "react-intl";
 import { Meta } from "typescript-clients/scim";
-import { getGroupDetails } from "../apis/scimGroupsRequest";
-import { deleteUser } from "../apis/scimUsersRequest";
+import { getGroupDetails } from "../apis/scim/groupsRequest";
+import { deleteUser } from "../apis/scim/usersRequest";
 import { fakePassword } from "../common/testEPPNData";
 import { useAppDispatch, useAppSelector } from "../hooks";
 import { getUsersSlice } from "../slices/getUsers";
